Consensus on these RERS examinations are wrong (sorry everyone !)
Due to ITS-Tools in 2020 believing NUPN implies one-safe, it answered
wrong to many queries, sometimes on its own, and sometime with benefit
of doubt (self-reinforced by two versions of my tool competing).
Anyway; these oracles cannot be trusted. afaik fortunately RERS is the
only model in the benchmark with NUPN that is not one safe.

Possibly some diagnosis on RERS in other categories are wrong, we only
tested RC and RF here.